Refactor API into routers and add Grafana dashboards
