+++ observability/grafana/provisioning/alerting/alerting.yml
@@ -1,0 +1,15 @@
+apiVersion: 1
+
+policies:
+  - orgId: 1
+    receiver: 'grafana-default-email'
+    group_by: ['grafana_folder', 'alertname']
+
+contactPoints:
+  - orgId: 1
+    name: 'grafana-default-email'
+    receivers:
+      - uid: 'default-email'
+        type: 'email'
+        settings:
+          addresses: 'admin@localhost'
